requirements.txt
langchain
langchain-core
langchain-anthropic
langgraph
duckduckgo-search
beautifulsoup4
reportlab
markdown2pdf
weasyprint
streamlit
python-dotenv
tavily-python
